refactor(ai): modularize AI decision pipeline into composable steps
Splits the monolithic AISimulator into discrete per-subsystem steps so
the AI can be extended without rewriting the whole pipeline:

- ai/steps/aiRouteStep.ts — route claiming / upkeep
- ai/steps/aiContractStep.ts — contract bid/accept logic
- ai/steps/aiHubStep.ts — hub build/upgrade
- ai/steps/aiTechStep.ts — research prioritization
- ai/steps/aiDecisionStep.ts — top-level orchestration + personality weighting

AISimulator now composes the steps rather than branching internally.
TurnSimulator and SimulationRunner call through the new pipeline, and
SimulationLogger grew structured fields for the new event types so the
AI sandbox feed can surface interactions.

All existing AI/sim integration tests continue to pass.
