+++ src/game/simulation/__tests__/Phase3Integration.test.ts
@@ -60,7 +60,7 @@ function makeGameState(overrides: Partial<GameState> = {}): GameState {
     companyName: "Test Corp",
     ceoName: "Commander",
     ceoPortrait: { portraitId: "ceo-01", category: "human" },
-    gameSize: "small",
+    gameSize: "standard",
     galaxyShape: "spiral",
     playerEmpireId: "emp-1",
     galaxy: {
@@ -168,6 +168,7 @@ function makeGameState(overrides: Partial<GameState> = {}): GameState {
       headwindBias: 0,
       turnsInDebt: 0,
       consecutiveProfitTurns: 0,
+      turnsSinceLastDecision: 0,
     },
     score: 0,
     gameOver: false,
@@ -185,6 +186,17 @@ function makeGameState(overrides: Partial<GameState> = {}): GameState {
     empireTradePolicies: {},
     interEmpireCargoLocks: [],
     stationHub: null,
+    saveVersion: 6,
+    actionPoints: { current: 2, max: 2 },
+    turnBrief: [],
+    pendingChoiceEvents: [],
+    activeEventChains: [],
+    captains: [],
+    routeMarket: [],
+    researchEvents: [],
+    unlockedNavTabs: ["map", "routes", "fleet", "finance"] as import("../../../data/types.ts").NavTabId[],
+    reputationTier: "unknown" as import("../../../data/types.ts").ReputationTier,
+    localRouteSlots: 2,
     ...overrides,
   };
 }
