+++ src/game/simulation/__tests__/SandboxSaveManager.test.ts
@@ -55,7 +55,7 @@ function createTestConfig(
 ): SimulationConfig {
   return {
     seed: 12345,
-    gameSize: "medium",
+    gameSize: "standard",
     galaxyShape: "spiral",
     companyCount: 6,
     maxTurns: 50,
@@ -76,7 +76,7 @@ function createTestGameState(overrides: Partial<GameState> = {}): GameState {
     companyName: "AI Sandbox Corp",
     ceoName: "Commander",
     ceoPortrait: { portraitId: "ceo-01", category: "human" },
-    gameSize: "medium",
+    gameSize: "standard",
     galaxyShape: "spiral",
     playerEmpireId: "",
     galaxy: { sectors: [], empires: [], systems: [], planets: [] },
@@ -91,6 +91,7 @@ function createTestGameState(overrides: Partial<GameState> = {}): GameState {
       headwindBias: 0,
       turnsInDebt: 0,
       consecutiveProfitTurns: 0,
+      turnsSinceLastDecision: 0,
     },
     score: 0,
     gameOver: false,
@@ -108,6 +109,17 @@ function createTestGameState(overrides: Partial<GameState> = {}): GameState {
     empireTradePolicies: {},
     interEmpireCargoLocks: [],
     stationHub: null,
+    saveVersion: 6,
+    actionPoints: { current: 2, max: 2 },
+    turnBrief: [],
+    pendingChoiceEvents: [],
+    activeEventChains: [],
+    captains: [],
+    routeMarket: [],
+    researchEvents: [],
+    unlockedNavTabs: ["map", "routes", "fleet", "finance"] as import("../../../data/types.ts").NavTabId[],
+    reputationTier: "unknown" as import("../../../data/types.ts").ReputationTier,
+    localRouteSlots: 2,
     ...overrides,
   };
 }
@@ -173,7 +185,7 @@ describe("SandboxSaveManager", () => {
       expect(data.meta.turn).toBe(5);
       expect(data.meta.maxTurns).toBe(50);
       expect(data.meta.status).toBe("running");
-      expect(data.meta.configSummary).toContain("Medium");
+      expect(data.meta.configSummary).toContain("Standard");
       expect(data.meta.configSummary).toContain("Spiral");
       expect(data.rngState).toBe(99);
       expect(data.speed).toBe("fast");
