+++ src/game/simulation/__tests__/TurnSimulator.test.ts
@@ -169,7 +169,7 @@ function makeGameState(overrides: Partial<GameState> = {}): GameState {
     companyName: "Test Corp",
     ceoName: "Commander",
     ceoPortrait: { portraitId: "ceo-01", category: "human" },
-    gameSize: "small",
+    gameSize: "standard",
     galaxyShape: "spiral",
     playerEmpireId: "",
     galaxy: {
@@ -193,6 +193,7 @@ function makeGameState(overrides: Partial<GameState> = {}): GameState {
       headwindBias: 0,
       turnsInDebt: 0,
       consecutiveProfitTurns: 0,
+      turnsSinceLastDecision: 0,
     },
     score: 0,
     gameOver: false,
@@ -210,6 +211,17 @@ function makeGameState(overrides: Partial<GameState> = {}): GameState {
     empireTradePolicies: {},
     interEmpireCargoLocks: [],
     stationHub: null,
+    saveVersion: 6,
+    actionPoints: { current: 2, max: 2 },
+    turnBrief: [],
+    pendingChoiceEvents: [],
+    activeEventChains: [],
+    captains: [],
+    routeMarket: [],
+    researchEvents: [],
+    unlockedNavTabs: ["map", "routes", "fleet", "finance"] as import("../../../data/types.ts").NavTabId[],
+    reputationTier: "unknown" as import("../../../data/types.ts").ReputationTier,
+    localRouteSlots: 2,
     ...overrides,
   };
 }
@@ -293,6 +305,7 @@ describe("TurnSimulator", () => {
           headwindBias: -0.5,
           turnsInDebt: 1, // Already 1 turn in debt
           consecutiveProfitTurns: 0,
+      turnsSinceLastDecision: 0,
         },
       });
